Revert "Don't apply channel modes received in an non-burst, non-creation FJOIN, because they could be out of date"
This keeps the protocol closer to TS6 behavior. The mode sync in the FJOIN is
used to resync channel modes after a desync caused by modes crossing on the
wire. In the future, an alternate join command will be used to prevent MODExJOIN
desync without rejecting modes from an FJOIN.

This reverts commit 4820e6bb9578e53ea1553070108599349d0ffbcb.

git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@11163 e03df62e-2008-0410-955e-edbf42e46eb7