(Hopefully) fix a issue with m_ident's result not being used for display.
The user class has a couple of caches for the user's hostmask. If one of those
caches is initialized before m_ident set the user's new ident it would stick
around until the cache was invalidated the next time. This patch fixes this by
using the proper API for setting a user's ident which also invalidates any
caches which might still have the old ident.


git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@11514 e03df62e-2008-0410-955e-edbf42e46eb7