Allow connect blocks to match the "servername" field in USER
RFC 1459 specifies that the third parameter to USER is the name of the
server that the client is connecting to. While this is not set by all
clients, it is set by some, and could be used to alter MOTD or other
settings based on the DNS name used to connect to a server or network.