Add maxconnwarn and show class in connectmsg
Created the maxconnwarn variable in the connect block, so you can make
connect blocks that only warns about max connections if you want to.
This reduces noise from connecting clients that have low maxlocal and/or
maxglobal.

In addition, made the connect announcement show the class a client is
in, making it easy to spot if a client is in the right class.