Don't throw an exception if passed bad parameters to stringjoiner.
It's far too common to "misuse" this API by accidentally trying to use
stringjoiner with an empty container, and fixing that is more or less adding
boilerplate in every place we fix it.

Instead, let's just return and not touch the string, and not throw.

Done-with: Attila Molnar <attilamolnar@hush.com>