Differ database error messages for m_permchannels and m_xline_db
m_permchannels and m_xline_db both have the same error messages as well as not
informing the server operator what file is at fault for the error. This makes it
hard to find which module is causing the error and what file the module is referring
to. The new messages explain which database is trying to be written and the path
used to write the file.