m_spanningtree Forward NICK messages when they cause a collision with the new nick param rewritten to the uuid
Suppressing these messages meant that we relied on an appropriate NICK to
come from the direction the SAVE was sent to; this left all servers behind
us uninformed until the target server of the SAVE reacted with a nick change
to uuid.

This was problematic because someone can legitimately change nick to the
past nick of the SAVEd user on a server that already has the SAVEd user
with the uuid nick and that nick change message can reach servers that
haven't yet seen the nick change to uuid of the SAVEd user.