Remove run-cc.pl and all associated code.
As far as I can tell this only worked with older versions of GCC
and there is no value in fixing it as diagnostics have improved
since then and it makes errors hard to search for anyway.

This also shaves off an average of five seconds on each build on
my hardware. Neat!