Fix sending malformed ERR_UNKNOWNCOMMAND messages in some cases.
This is not something the average user will encounter. It can only
happen if the user sends a message with preceding whitespace or a
prefix but no command name.

This is not something that should ever be seen in practise so we
just penalise the user and pretend nothing ever happened.

The previous code also contained undefined behaviour but it acted
sensibly on all compilers we support so it was not crashable.