Change the branch version from InspIRCd-3.0 to InspIRCd-3.
Under the new versioning scheme when a new branch is forked the
major version number will be reset. The minor version number will
now be used to mark a release as containing backwards compatible
additions and the patch version number will be used to mark a
release as containing just bug fixes.