Fix not propagating rehashes properly across the network.
The REHASH command is not routed like a normal command. When a
rehash that needs to be forwarded is received the local server
forwards it indirectly by calling the local REHASH handler which
then calls the OnPreRehash hook. This results in the OnPreRehash
hook being called whilst loopCall is true.

The bug which this check was added to fix seems to no longer be
present so no alternate fix for that seems necessary.

Closes #1537.