DurationString(): Return "0s" for a duration of 0 (#1677).
Usually a duration of 0 is not allowed or handled separately,
but it can also be used as a 'no set time' without separation.

Case in point: m_chanhistory calls DurationString() to convert
the max time seconds back to a human readable string for the
mode serializer. Returning a blank string is bad here.