Improve handling of a link block without a port.
Currently if you define a link block without a port for an inbound
only link (e.g., services), the allowmasks and/or address are not
added to the 'ValidIPs' vector. This will cause a denial when the
connection is starting unless you have another link block (with a
port) that includes the same address (or matching allowmask).
You can also attempt to /connect a server with no defined port, it
will fail to connect instantly but doesn't need to even try.

Removing the 'no port' check changes this behaviour to better
allow link blocks with no port defined. Adding a 'no port' check
to the /connect command handler prevents the attempt at trying to
connect to one of these links, with an error message as to why.