Change Config->CaseMapping back when unloading.
Now that casemapping is configurable in the core and we set that
Config variable rather than just modifying the ISupport output each
time, we need to change the variable back when being unloaded. So we
save the current value when loading and set it back when being
unloaded. We also need to call the ISupport builder a second time as
the core calls it before we destruct.