Don't allow routing INFO to remote servers.
There's no reason to request this for a remote server. It rarely
differs between InspIRCd versions and neither Anope nor Atheme
implement handlers for it.

We don't need to implement anything to the protocol compat layer
for this as if v3 servers send an INFO request to a v4 server they
they will just be ignored as the command has no HandleRemote impl.

Closes #1943.