Fix our WHO oplevels being incompatible with the WHOX spec.
Unlike in InspIRCd where 0 is an unprivileged member and higher
levels grant more privileges the WHOX spec defines the oplevel
field as using n/a for unprivileged users and the lower an oplevel
is the more privileges it grants. InspIRCd also uses values between
0 and UINT_MAX whereas WHOX only uses values between 0 and 999.

In order to work around this we now lazily build dummy oplevels for
the InspIRCd member ranks.