Use consistent xline messages locally
For connectban these messages would have previously been duplicated
to other servers until this was fixed earlier today. For dnsbl and
rline they would have just differed on the local to a remote.

As of the previous commit the module name is included in the setter
so there is no need to vary the snotice. This makes things more
consistent for server operators and allows scripts to parse the
messages.