Work around a bug in reading the config for the Argon2 module.
ProviderConfig::version is set to an argon2 version not a base 10
version so using it as a default for the config may result in the
default being ignored.

Luckily 0x10 and 0x12 are 16 and 18 respectively so we can just
allow them as valid values until v4 where we can eradicate the
SanitiseArgon2Version function and use getEnum instead.