Allow secondary cloak methods to be used on link synchronisation.
If a server has a conditional cloak method (e.g. account) as the
primary cloak method it doesn't make sense to use it for link
consistency. With this change a secondary cloak method can request
that it is used instead of the primary if the primary does not
mark itself as link sensitive.