Improve the consistency of duration strings.
Unfortunately because of leap years we can't really make duration
strings exact. Because of this we have always added 6 hours on to
each year to make it correct over time.

However, we did not do this in any kind of consistent way which
resulted in weird roundtripped duration strings which made this
very confusing for users

While I'm touching the duration code I've also cleaned it up.